ex00 complete

+++ cpp08/ex00/easyfind.hpp
@@ -10,8 +10,8 @@ public:
 };
 
 template<typename T>
-int easyfind(const T& container, int value) {
-    typename T::const_iterator it = std::find(container.begin(), container.end(), value);
+long easyfind(const T& container, int value) {
+    typename T::const_iterator it = std::find(container.begin(), container.end(), value); // tell compiler T is a type
     if (it == container.end()) {
         throw NotFoundException();
     }
